Implement stage registry for visual effects
